Rendering audio files
Added RoatationinSeconds.scd which has a version of the rotation model
for 3channels where 1 earth orbit is 36.5 seconds. The planet excerpt
files follow this form. There is one version that in the minute format
for all planets (mercury through jupiter) where 1 earth orbit is 365
days. The 3 channel output creates a phantom speaker behind the
listener so it may “disappear” in headphones but creates a better
stereo image of rotation.